feat: add support for environment variables
Add support for environment variables in .env and .env.example files.
This allows some configuration settings to be set at runtime, which can be useful for quickly changing the configuration of the application without having to rebuild it.

BREAKING CHANGE: The URL of the Swagger UI has been changed to reflect the new environment variable settings.